requirements.txt
pymongo==3.12.0
Flask==2.0.1
flask_swagger_ui==3.36.0
uwsgi
paho-mqtt
requests
matplotlib
tk